create goto admin middleware

+++ se_blog/routes/web.php
@@ -25,23 +25,23 @@ Route::get('/blog/list','BlogController@getList')->name('blog.list');
 
 Route::get('/blog/create',function (){
     return view('blog.create');
-})->name('blog.create');
+})->middleware('admin')->name('blog.create');
 
-Route::get('/blog/update/{id}','BlogController@toUpdate');
+Route::get('/blog/update/{id}','BlogController@toUpdate')->middleware('admin');
 
 
 Route::get('/blog/{id}', 'BlogController@getOne')->middleware('auth')->name('blog');
 
-Route::post('/blog','BlogController@create')->name('blog');
+Route::post('/blog','BlogController@create')->middleware('admin')->name('blog');
 
-Route::patch('/blog/{id}', 'BlogController@update');
+Route::patch('/blog/{id}', 'BlogController@update')->middleware('admin');
 
-Route::delete('/blog/{id}','BlogController@destroy');
+Route::delete('/blog/{id}','BlogController@destroy')->middleware('admin');
 
 
-Route::get('admin/report','ExportExcelController@index');
-Route::get('/report/download','ExportExcelController@excel')->name('export_excel.excel');
-Route::get('/report/blogs/download','ExportExcelController@excelBlogs')->name('export_blogs.excel');
+Route::get('admin/report','ExportExcelController@index')->middleware('admin');
+Route::get('/report/download','ExportExcelController@excel')->middleware('admin')->name('export_excel.excel');
+Route::get('/report/blogs/download','ExportExcelController@excelBlogs')->middleware('admin')->name('export_blogs.excel');
 
 
 // OAuth Routes
